Move class-related classes to Classes namespace
