+++ tests/source/Mocka/ClassTraitTest.php
@@ -9,19 +9,19 @@ class ClassTraitTest extends \PHPUnit_Framework_TestCase {
     public function testMockClass() {
         $mocka = new Mocka();
         $mockClass = $mocka->mockClass('\MockaMocks\AbstractClass');
-        $this->assertInstanceOf('\\Mocka\\ClassMock', $mockClass);
+        $this->assertInstanceOf('\\Mocka\\Classes\\ClassMock', $mockClass);
     }
 
     public function testMockInterface() {
         $mocka = new Mocka();
         $mockClass = $mocka->mockInterface('\MockaMocks\InterfaceMock');
-        $this->assertInstanceOf('\\Mocka\\ClassMock', $mockClass);
+        $this->assertInstanceOf('\\Mocka\\Classes\\ClassMock', $mockClass);
     }
 
     public function testCallMockedMethod() {
         $mocka = new Mocka();
         $mockClass = $mocka->mockClass('\MockaMocks\AbstractClass');
-        /** @var \MockaMocks\AbstractClass|\Mocka\ClassMockTrait $object */
+        /** @var \MockaMocks\AbstractClass|\Mocka\Classes\ClassMockTrait $object */
         $object = $mockClass->newInstanceWithoutConstructor();
         $this->assertNull($object->foo());
         $this->assertSame('bar', $object->bar());
